+++ tests/util.ts
@@ -33,14 +33,3 @@ export const copyFile = (
         resolve();
       });
   });
-
-export const promiseRimraf = async (path: string): Promise<boolean> =>
-  new Promise((resolve) => {
-    try {
-      rimraf(path, () => {
-        resolve(true);
-      });
-    } catch (e) {
-      resolve(false);
-    }
-  });